pyproject.toml
[tool.poetry]
name = "world_state"
version = "0.1.0"
description = ""
authors = ["Jonathan Leeming <jonathan@leeming.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.11"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
